Rename components to match refactor in /show_item
`/show_item` branch used this branch (`/all_items`) as a base.  We are renaming here in hopes of reducing the amount of merge conflicts that will occur when `/show_item` needs to rebase this branch.
