+++ src/components/all_items_page/ItemsPage.scss
