+++ src/components/item_card/ItemCard.scss
